more spec and class cleanup to get pdk unit tests working
