+++ spec/type_aliases/dns_record_aliases_spec.rb
@@ -1,0 +1,286 @@
+require 'spec_helper'
+
+describe 'dns::record::aaaa', type: :define do
+  let(:pre_condition) { 'include ::dns::server' }
+  let :facts do
+    {
+      concat_basedir: '/tmp',
+    }
+  end
+
+  context 'letting the host be defined by the resource name' do
+    let(:title) { 'foo' }
+    let :params do
+      {
+        zone: 'example.com',
+        data: ['::1'],
+      }
+    end
+
+    it { is_expected.not_to raise_error }
+    it {
+      is_expected.to contain_concat__fragment('db.example.com.foo,AAAA,example.com.record')
+        .with_content(%r{^foo\s+IN\s+AAAA\s+::1$})
+    }
+  end
+  context 'assigning a different host than the resource name' do
+    let(:title) { 'foo' }
+    let :params do
+      {
+        zone: 'example.com',
+        host: 'bar',
+        data: ['::1'],
+      }
+    end
+
+    it { is_expected.not_to raise_error }
+    it {
+      is_expected.to contain_concat__fragment('db.example.com.foo,AAAA,example.com.record')
+        .with_content(%r{^bar\s+IN\s+AAAA\s+::1$})
+    }
+  end
+end
+describe 'dns::record::a', type: :define do
+  let(:pre_condition) { 'include ::dns::server' }
+  let :facts do
+    {
+      concat_basedir: '/tmp',
+    }
+  end
+
+  context 'letting the host be defined by the resource name' do
+    let(:title) { 'foo' }
+    let :params do
+      {
+        zone: 'example.com',
+        data: ['1.2.3.4'],
+      }
+    end
+
+    it { is_expected.not_to raise_error }
+    it {
+      is_expected.to contain_concat__fragment('db.example.com.foo,A,example.com.record')
+        .with_content(%r{^foo\s+IN\s+A\s+1\.2\.3\.4$})
+    }
+  end
+  context 'assigning a different host than the resource name' do
+    let(:title) { 'foo' }
+    let :params do
+      {
+        zone: 'example.com',
+        host: 'bar',
+        data: ['1.2.3.4'],
+      }
+    end
+
+    it { is_expected.not_to raise_error }
+    it {
+      is_expected.to contain_concat__fragment('db.example.com.foo,A,example.com.record')
+        .with_content(%r{^bar\s+IN\s+A\s+1\.2\.3\.4$})
+    }
+  end
+end
+describe 'dns::record::cname', type: :define do
+  let(:pre_condition) { 'include ::dns::server' }
+  let :facts do
+    {
+      concat_basedir: '/tmp',
+    }
+  end
+
+  context 'letting the host be defined by the resource name' do
+    let(:title) { 'foo' }
+    let :params do
+      {
+        zone: 'example.com',
+        data: 'baz.example.com',
+      }
+    end
+
+    it { is_expected.not_to raise_error }
+    it {
+      is_expected.to contain_concat__fragment('db.example.com.foo,CNAME,example.com.record')
+        .with_content(%r{^foo\s+IN\s+CNAME\s+baz\.example\.com\.$})
+    }
+  end
+  context 'assigning a different host than the resource name' do
+    let(:title) { 'foo' }
+    let :params do
+      {
+        zone: 'example.com',
+        host: 'bar',
+        data: 'baz.example.com',
+      }
+    end
+
+    it { is_expected.not_to raise_error }
+    it {
+      is_expected.to contain_concat__fragment('db.example.com.foo,CNAME,example.com.record')
+        .with_content(%r{^bar\s+IN\s+CNAME\s+baz\.example\.com\.$})
+    }
+  end
+end
+describe 'dns::record::mx', type: :define do
+  let(:pre_condition) { 'include ::dns::server' }
+  let :facts do
+    {
+      concat_basedir: '/tmp',
+    }
+  end
+
+  context 'letting the host be defined by the resource name' do
+    let(:title) { 'foo' }
+    let :params do
+      {
+        zone: 'example.com',
+        data: 'baz.example.com',
+        preference: 10,
+      }
+    end
+
+    it { is_expected.not_to raise_error }
+    it {
+      is_expected.to contain_concat__fragment('db.example.com.foo,MX,example.com.record')
+        .with_content(%r{^foo\s+IN\s+MX\s+10\s+baz\.example\.com\.$})
+    }
+  end
+  context 'assigning a different host than the resource name' do
+    let(:title) { 'foo' }
+    let :params do
+      {
+        zone: 'example.com',
+        host: 'bar',
+        data: 'baz.example.com',
+        preference: 10,
+      }
+    end
+
+    it { is_expected.not_to raise_error }
+    it {
+      is_expected.to contain_concat__fragment('db.example.com.foo,MX,example.com.record')
+        .with_content(%r{^bar\s+IN\s+MX\s+10\s+baz\.example\.com\.$})
+    }
+  end
+end
+describe 'dns::record::ns', type: :define do
+  let(:pre_condition) { 'include ::dns::server' }
+  let :facts do
+    {
+      concat_basedir: '/tmp',
+    }
+  end
+
+  context 'letting the host be defined by the resource name' do
+    let(:title) { 'foo' }
+    let :params do
+      {
+        zone: 'example.com',
+        data: 'baz.example.com.',
+      }
+    end
+
+    it { is_expected.not_to raise_error }
+    it {
+      is_expected.to contain_concat__fragment('db.example.com.foo,NS,example.com.record')
+        .with_content(%r{^foo\s+IN\s+NS\s+baz\.example\.com\.$})
+    }
+  end
+  context 'assigning a different host than the resource name' do
+    let(:title) { 'foo' }
+    let :params do
+      {
+        zone: 'example.com',
+        host: 'bar',
+        data: 'baz.example.com.',
+      }
+    end
+
+    it { is_expected.not_to raise_error }
+    it {
+      is_expected.to contain_concat__fragment('db.example.com.foo,NS,example.com.record')
+        .with_content(%r{^bar\s+IN\s+NS\s+baz\.example\.com\.$})
+    }
+  end
+end
+describe 'dns::record::ptr', type: :define do
+  let(:pre_condition) { 'include ::dns::server' }
+  let :facts do
+    {
+      concat_basedir: '/tmp',
+    }
+  end
+
+  context 'letting the host be defined by the resource name' do
+    let(:title) { '1' }
+    let :params do
+      {
+        zone: '0.0.127.in-addr.arpa',
+        data: 'localhost',
+      }
+    end
+
+    it { is_expected.not_to raise_error }
+    it {
+      is_expected.to contain_concat__fragment('db.0.0.127.in-addr.arpa.1,PTR,0.0.127.in-addr.arpa.record')
+        .with_content(%r{^1\s+IN\s+PTR\s+localhost\.$})
+    }
+  end
+  context 'assigning a different host than the resource name' do
+    let(:title) { 'foo' }
+    let :params do
+      {
+        zone: '0.0.127.in-addr.arpa',
+        host: '1',
+        data: 'localhost',
+      }
+    end
+
+    it { is_expected.not_to raise_error }
+    it {
+      is_expected.to contain_concat__fragment('db.0.0.127.in-addr.arpa.foo,PTR,0.0.127.in-addr.arpa.record')
+        .with_content(%r{^1\s+IN\s+PTR\s+localhost\.$})
+    }
+  end
+end
+
+describe 'dns::record::txt', type: :define do
+  let(:pre_condition) { 'include ::dns::server' }
+  let :facts do
+    {
+      concat_basedir: '/tmp',
+    }
+  end
+
+  context 'letting the host be defined by the resource name' do
+    let(:title) { 'foo' }
+    let :params do
+      {
+        zone: 'example.com',
+        data: 'baz',
+      }
+    end
+
+    it { is_expected.not_to raise_error }
+    it {
+      is_expected.to contain_concat__fragment('db.example.com.foo,TXT,example.com.record')
+        .with_content(%r{^foo\s+IN\s+TXT\s+"baz"$})
+    }
+  end
+
+  context 'assigning a different host than the resource name' do
+    let(:title) { 'foo' }
+    let :params do
+      {
+        zone: 'example.com',
+        host: 'bar',
+        data: 'baz.example.com',
+      }
+    end
+
+    it { is_expected.not_to raise_error }
+    it {
+      is_expected.to contain_concat__fragment('db.example.com.foo,TXT,example.com.record')
+        .with_content(%r{^bar\s+IN\s+TXT\s+"baz"$})
+    }
+  end
+end
